+++ extensions/audio/src/main/java/ai/djl/audio/processor/package-info.java
@@ -11,11 +11,5 @@
  * and limitations under the License.
  */
 
-package ai.djl.audio.featurizer;
-
-import ai.djl.ndarray.NDArray;
-import ai.djl.ndarray.NDManager;
-
-public interface AudioProcessor {
-    NDArray extractFeatures(NDManager manager, NDArray samples);
-}
+/** Contains multiple processors for audio and signal processing */
+package ai.djl.audio.processor;
